TEST:
Forcing the value of DotNetCliVersion to be empty to see how the system behaves if valid package type is set or there is no known version information for the package type parameter is passed in.

The expectation is having a empty value will result in some kind of null reference error when trying to access .Net Core SDK data.

If the value does not control the actual SDK install and or use it will also let us know if the version test script that is failing is doing a simple string comparison as opposed to examining the actual SDK being used. If it IS doing a simple string comparison the expectation is that the output of the failure will not show an empty value for the current .Net CORE SDK in the mismatch error rather than SDK version that is actually installed.