Add Artifact repo extension point to Sign and Publish
When repositories don't want to publish the default set of artifacts
that are globbed, they need to align the items that are (post build)
signed and published. To make easily possible, add an `Artifact` item
extension point that is respected in both Sign.proj and Publish.proj.

Also rename the previous EnableDefaultPublishItems to EnableDefaultItems
as it now applies to both signing and publishing.

Import Sign.props earlier so that Publishing.props (repo file) can
use its properties and items outside of targets.