Make init-compiler.sh fall back to and work with generic compiler
If a version-specific compiler (eg, clang-17) is passed to the script,
it works as before. And like before, if a generic compiler (eg clang) is
passed to the script, it prefers to find and use the few
specific/known/tested versions of compilers (eg, clang means use
clang-18 if possible). However, unlike before, if none of those specific
known versions are found, it correctly falls back to the generic
compiler binary (eg clang) in $PATH.

The script was actually supposed to do all this already, but it seems to
have been broken when warning messages were added. These warning
messages were parsed by msbuild. Msbuild would see the warning messages
and fail the build.

With a generic compiler, this script now also invokes it via $compiler
-dumpversion (which works on both clang and gcc) to find the version for
various (existing) version checks. The compiler was already being
invoked to check if -fuse-ld=lld is supported.

Fixes: #14632