Disable central package management in Tools.csproj
Having Directory.Packages.props imported automatically is problematic
because this project is not setup for working with central package
management. Auto importing it makes it **very** difficult for repos that
use CPM to work with Tools.csproj.

Until this project is made to work with CPM it shouldn't be importing
Directory.Packages.props

closes #14693