Add template to generate a delegation SAS token for storage This template uses an MI to log in and generate a delegation SAS token for a storage account + container. When replacing storage account and account key SAS usage, we have a few options: - Direct use of an MI from the environment (logged in or on the machine) via BlobServiceClient in code. - An ADO account token for storage, used as a bearer token in authentication headers - Delegation SAS.
The most desirable option is direct use of an MI. But this option has some problems in build scenarios. .NET builds use dotnet-install to install internal bits from storage. dotnet-install doesn't have a way to support managed identities at this point. Furthermore, it doesn't really even know that it's talking to a storage account (it's just contructing a URI and downloading from it) and may not be talking to a storage account in some use cases. Direct support for managed identities is likely to be a bit hacky in the current script design. We could use an account token in an auth header, but this would require changes in dotnet-install (possible and being discussed) and also doesn't work with aka.ms links. Furthermore, the hour long token lifetime would present a significant problem in long-running build or test scenarios. The last option has the benefit of being able to be used with existing infra and the token lifetime can be set according to needs.

This template does the following:
- Uses the AzureCLI + federated service connection to generate a delegation SAS token for a storage account + container.
- The token is stored in the specified variable
- The token is valid for 1 hour by default. The maximum lifetime is 7 days.
- By default the perms are read and list.
- The token can be encoded in base64 to make passing through various scripting languages easier.

At this point, I don't think that it's necessary to support different versions for *Nix and Windows. Powershell will be available (or can be added) in all the use cases.