Remove WPF work around
This WPF work around is breaking incremental builds in Visual Studio.
The `<BuildingProject>false</BuildingProject>` explicit setting
triggers the target `_ComputeNonExistentFileProperty` which breaks up to
date checks by design. This is showing up as a reason for Roslyn builds
to not be incremental in VS.

The associated issue in the project system's been closed for four years
now. This should be safe to remove or if it's not we should be
reactivating the issue in project-system.