Support building sid and alpineedge rootfs for loongarch64
This is a step towards eventual access to public CI resources to catch
build regressions etc. earlier. As LoongArch support has not reached any
public release of Debian or Alpine, currently only sid and alpineedge
are covered.

This is tested on a x86_64 system with qemu-loongarch64 linux-user
emulation set up, because the debootstrap second stage needs to run
binaries of the target architecture. The following invocations work:

* `eng/common/cross/build-rootfs.sh loongarch64 sid --rootfsdir /test`
* `eng/common/cross/build-rootfs.sh loongarch64 alpineedge --skipsigcheck --rootfsdir /test`