Support building sid and alpine{,edge} rootfs for loongarch64
This is a step towards eventual access to public CI resources to catch
build regressions etc. earlier. As LoongArch support has not reached any
public release of Debian, currently only sid is covered, while alpine
versions from 3.21 onwards are supported.

This is tested on a x86_64 system with qemu-loongarch64 linux-user
emulation set up, because the debootstrap second stage needs to run
binaries of the target architecture. The following invocations work:

* `eng/common/cross/build-rootfs.sh loongarch64 sid --rootfsdir /test`
* `eng/common/cross/build-rootfs.sh loongarch64 alpineedge --skipsigcheck --rootfsdir /test`
* `eng/common/cross/build-rootfs.sh loongarch64 alpine --skipsigcheck --rootfsdir /test`