Replace strong name verification with custom implementation The goal here is two-fold: - Strong name checking could run on non-Windows platforms (not be reliant on sn.exe) - If a file has a valid strong name, do not strong name it.
The implementation was pulled from a combination of roslyn's strong naming tests and runtime's checksum implemention and a few constants. The implementation was altered to remove need for private reflection and avoid unsafe code when calculating checksums. When a PE file is found, we check whether it has a valid strong name. If it does, and we would have tried to strong name sign it, avoid doing so.

This should allow Mac and Linux machines to consume binaries from early build stages and run signtool without attempting to re-strong name.