Make AssetManifest* properties more extensible and use a better default in CI
Move the AssetManifestFileName and AssetManifestFilePath properties after the Publishing.props import so that repos don't have to redefine `AssetManifestFilePath` anymore when changing `AssetManifestFileName`.

When `AssetManifestFileName` isn't provided externally, use the Agent.JobName env var when available. That is guaranteed to be unique and avoids the need for repos to come up with a unique name themselves. This should remove the need for all cases that pass a name in externally.