Opt-in TargetRid publish mode
To enable us to fully dedup our vertical asset manifests in the VMR, we need some mechanism to specify that RID-agnostic packages shouldn't be published in particular verticals.

One options would be to have the VMR automatically use "Vertical" visibility as the default for all non-"Win-x64 BuildPass1" legs and require repos that have RID-specific assets to mark them with the correct visibility. In this model, every repo that has RID specific assets would have to have almost the same logic in Publishing.props.

Another option (this PR) is to add support into Arcade for a switch to automatically only add RID-specific packages as non-Vertical packages by default. This model ensures that we don't need to change every repo that produces RID-specific artifacts. It also provides a mechanism to opt-in for repositories that need to publish in this manner normally (useful for dotnet/runtime for the short period between the MicroBuild fixes and the official build being shut off).