Copy Maestro tasks back into Arcade and update to new APIs Copies the Maestro tasks back into arcade and updates 10.0 Arcade to keep using those tasks.
The eventual goal here is that most likely all the publishing and manifest management will move out of arcade and into a separate repo which serves all of the repos. We're a ways away from this though, as it means the manifest handling and task need to be compatible with all versions of .NET (this is no longer the case as signing data has been removed). For now, I believe the right way to go is to consider the task dead in arcade-services for .NET 10+, but keep it around for fixes if necessary (unlikely). Make tweaks as necessary for new publishing versions in .NET 10 arcade. When .NET 6 is well and truly gone, move the publishing and manifest code out of arcade and into a new repo with a design focused on preserving support for old manifest+publishing versions. The arcade services code would get deleted at this time.