Add artifact visibility support to GenerateBuildManifest
This allows us to use GenerateBuildManifest to split the "join verticals" job in the VMR and have a "build the merged manifest" job and a "download/upload artifacts" job.

Alternatively, we can remove GenerateBuildManifest (it has no users today) and add a "don't publish artifacts" option to PushToBuildStorage so it just produces the build model.