Provide an injection point for custom test reporting functionality
The current test logging implementation makes it difficult to know what
tests failed during test execution.
This especially affects GitHub Actions, which lack any test reporting
functionality, and require developers to download logs and search
for failures.

The injection point provides the ability to mitigate the issue by
allowing to inject a custom test reporter project before a test run
is terminated (in an event of a failure).