Follow Deterministic configuration when using NuGetPack
This should be a no-op because PackageBuilder currently ignores the
setting. However, if/when PackageBuilder starts following it, we should
start getting deterministic builds too.

Also renamed the Task invocation in the targets file to make the casing
of the task definition. It doesn't matter because MSBuild is
case-insensitive, but it's easier to search for.