Add ability to opt-out of ESRP service connection use.
I chose to use microbuildUseESRP rather than microbuildSignType because I don't want users to confuse use of the parameter with setting signing to test signing or real signing. _SignType is so prolific that it's hard to imagine getting rid of it at this point. So I chose just to have a parameter that indicates whether the service connection should be used. A user would pass false in test signing cases. This could be passed based on the value of _SignType if it's defined in the same file.