Refactor vsRequirements assignment logic in tools.ps1
Because 17.14 went stable, we needed a way to run vswhere, find the VS node, and enable preview SDKs. Noah found a way to do this but he had to modify tools.ps1 because the SDK repo didn't have a VS node in our global.json

https://github.com/dotnet/sdk/pull/51558/files#diff-72b8f8e899b94872c6ead31fd06ec109da15bcb9ad2af6e78103d6763a31c637

Porting his change here to see if folks want this centrally. Alternatively, I'm trying adding the vs node to global.json to see if it unblocks us.